Had to make sure all the rfm22 merge uses uintptr_t correctly

Conflicts:
	flight/Modules/Radio/radio.c
	flight/PiOS/Boards/STM32103CB_PIPXTREME_Rev1.h
	flight/PiOS/Boards/STM32F4xx_Freedom.h
	flight/PiOS/Boards/STM32F4xx_RevoMini.h
	flight/PiOS/inc/pios_com.h
	flight/targets/Freedom/System/pios_board.c
	flight/targets/PipXtreme/System/pios_board.c
	flight/targets/RevoMini/System/pios_board.c

+++ flight/PiOS/Boards/STM32103CB_PIPXTREME_Rev1.h
@@ -172,13 +172,13 @@ extern uint32_t pios_i2c_flexi_adapter_id;
 //-------------------------
 #define PIOS_COM_MAX_DEVS			5
 
-extern uint32_t pios_com_telem_usb_id;
-extern uint32_t pios_com_telem_vcp_id;
-extern uint32_t pios_com_telem_uart_telem_id;
-extern uint32_t pios_com_telem_uart_flexi_id;
-extern uint32_t pios_com_telemetry_id;
-extern uint32_t pios_com_rfm22b_id;
-extern uint32_t pios_com_radio_id;
+extern uintptr_t pios_com_telem_usb_id;
+extern uintptr_t pios_com_telem_vcp_id;
+extern uintptr_t pios_com_telem_uart_telem_id;
+extern uintptr_t pios_com_telem_uart_flexi_id;
+extern uintptr_t pios_com_telemetry_id;
+extern uintptr_t pios_com_rfm22b_id;
+extern uintptr_t pios_com_radio_id;
 extern uint32_t pios_ppm_rcvr_id;
 #define PIOS_COM_TELEM_USB         (pios_com_telem_usb_id)
 #define PIOS_COM_TELEM_VCP         (pios_com_telem_vcp_id)
