+++ flight/PiOS/Boards/STM32F30x_FlyingF3.h
@@ -137,11 +137,11 @@ extern uint32_t pios_i2c_external_id;
 // See also pios_board.c
 //-------------------------
 #define PIOS_COM_MAX_DEVS               4
-extern uint32_t pios_com_telem_rf_id;
-extern uint32_t pios_com_gps_id;
-extern uint32_t pios_com_telem_usb_id;
-extern uint32_t pios_com_bridge_id;
-extern uint32_t pios_com_vcp_id;
+extern uintptr_t pios_com_telem_rf_id;
+extern uintptr_t pios_com_gps_id;
+extern uintptr_t pios_com_telem_usb_id;
+extern uintptr_t pios_com_bridge_id;
+extern uintptr_t pios_com_vcp_id;
 #define PIOS_COM_GPS                    (pios_com_gps_id)
 #define PIOS_COM_TELEM_USB              (pios_com_telem_usb_id)
 #define PIOS_COM_TELEM_RF               (pios_com_telem_rf_id)
@@ -149,7 +149,7 @@ extern uint32_t pios_com_vcp_id;
 #define PIOS_COM_VCP                    (pios_com_vcp_id)
 
 #if defined(PIOS_INCLUDE_DEBUG_CONSOLE)
-extern uint32_t pios_com_debug_id;
+extern uintptr_t pios_com_debug_id;
 #define PIOS_COM_DEBUG                  (pios_com_debug_id)
 #endif	/* PIOS_INCLUDE_DEBUG_CONSOLE */
