+++ flight/PiOS/Boards/STM32F4xx_OSD.h
@@ -178,10 +178,10 @@ extern uint32_t pios_spi_port_id;
 // See also pios_board.c
 //-------------------------
 #define PIOS_COM_MAX_DEVS               5
-extern uint32_t pios_com_telem_rf_id;
-extern uint32_t pios_com_gps_id;
-extern uint32_t pios_com_aux_id;
-extern uint32_t pios_com_telem_usb_id;
+extern uintptr_t pios_com_telem_rf_id;
+extern uintptr_t pios_com_gps_id;
+extern uintptr_t pios_com_aux_id;
+extern uintptr_t pios_com_telem_usb_id;
 #define PIOS_COM_AUX                    (pios_com_aux_id)
 #define PIOS_COM_GPS                    (pios_com_gps_id)
 #define PIOS_COM_TELEM_USB              (pios_com_telem_usb_id)
@@ -198,7 +198,7 @@ extern uint32_t pios_com_telem_usb_id;
 
 
 #if defined(PIOS_INCLUDE_USB_HID)
-extern uint32_t pios_com_telem_usb_id;
+extern uintptr_t pios_com_telem_usb_id;
 #define PIOS_COM_TELEM_USB              (pios_com_telem_usb_id)
 #endif
